various changes to clean up code - removed unused lines of code in c-lib - renamed dataclasses.py to it_dataclasses.py to avoid conflict
